Rename dirs back (package name can't be changed in GooglePlay)
