+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.luckywarepro.musicintervals2anki" >
+    package="com.ichi2.apisample" >
+    <!-- should be changed to package="com.luckywarepro.musicintervals2anki" -->
 
     <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />
