+++ app/src/main/res/layout/activity_main.xml
@@ -6,7 +6,7 @@
                 android:paddingRight="@dimen/activity_horizontal_margin"
                 android:paddingTop="@dimen/activity_vertical_margin"
                 android:paddingBottom="@dimen/activity_vertical_margin"
-                tools:context="com.luckywarepro.musicintervals2anki.MainActivity">
+                tools:context="com.luckywarepro.apisample.MainActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
