+++ app/src/test/java/com/ichi2/apisample/AnkiDroidHelperTest.java
@@ -1,4 +1,4 @@
-package com.luckywarepro.musicintervals2anki;
+package com.ichi2.apisample;
 
 import org.junit.Test;
